Prolog BB m4: fix unsealed-RIP lea bombs in bb_list.cpp + bb_is_cmp.cpp; extend gz_arith_const_eval (bitwise /\ \/ xor >> << max min gcd ** sign truncate integer); add gz_arith_float_eval + rt_pl_is_cell_float for float-const is/2 (pi exp sqrt sin cos etc); fix x86_movimm binary sign-extension (uint32->uint64). m3 91/115 (floor held), m4 75->87 (+12). GATE-1 5/5/5 SNO 7/7/7 Icon 12/10/10 invariant.

+++ src/emitter/BB_templates/x86_asm.h
@@ -173,7 +173,7 @@ inline std::string x86_pop (const char * r) { int m = x86_rnum(r); std::string c
 /*--------------------------------------------------------------------------------------------------------------------*/
 inline std::string x86_movimm(const char * dst, long imm) {
     int m = x86_rnum(dst);
-    if (MEDIUM_BINARY) { std::string code; uint8_t rex = 0x48; if (m >= 8) rex |= 0x01; code += (char)rex; code += (char)(0xB8 | (m & 7)); code += u64le((uint64_t)(uint32_t)imm); return x86_Lrec(code); }
+    if (MEDIUM_BINARY) { std::string code; uint8_t rex = 0x48; if (m >= 8) rex |= 0x01; code += (char)rex; code += (char)(0xB8 | (m & 7)); code += u64le((uint64_t)imm); return x86_Lrec(code); }
     return std::string(" mov ") + dst + ", " + std::to_string(imm) + "\n";
 }
 inline std::string x86_movimm32(const char * dst, long imm) {
